+++ WebApplication/src/store/modules/i18n.js
@@ -5,17 +5,10 @@ let allLanguages = [
     { name: "Français", key: "fr" },
     { name: "Italiano", key: "it" },
     { name: "ქართული", key: "ka" },
+    { name: "Română", key: "ro" },
     { name: "Русский", key: "ru" },
     { name: "Slovensko", key: "si" },
     { name: "Українська", key: "uk" },
     { name: '中文', key: 'zh' }
-    // { name: 'Português', key: 'pt' },
-    // { name: 'Deutsche', key: 'de' },
-    // { name: 'Nederlands', key: 'nl' },
-    // { name: '日本語', key: 'jp' },
-    // { name: 'বাংলা', key: 'bn' },
-    // { name: 'Bahasa', key: 'xn' },
-    // { name: 'हिंदू', key: 'hi' },
-    // { name: 'Kiswahili', key: 'sw' },
 ]
 export default allLanguages;