+++ release-notes.md
@@ -2,6 +2,8 @@ Pending changes (not released yet)
 ==================================
 
 Changes:
+  - New configuration "UiOptions.DefaultOrdering" and new default value for 
+    "StudyListContentIfNoSearch": "default-ordering".
   - Added Romanian translations.