+++ release-notes.md
@@ -7,6 +7,10 @@ Changes:
   - Now showing the status of the Housekeeper plugin in the "System Info"
   - Added Romanian translations.
 
+Fixes:
+  - Fix [#72](https://github.com/orthanc-server/orthanc-explorer-2/issues/72): 
+    Can't modify StudyDescription if PatientName and PatientBirthDate are empty and if Patient has multiple studies
+
 
 1.8.1 (2025-04-01)
 ==================